fix tests for generic constraint check

+++ _test/gen8.go
@@ -7,6 +7,9 @@ type Float interface {
 func add[T Float](a, b T) float64 { return float64(a) + float64(b) }
 
 func main() {
-	var x, y float64 = 1, 2
+	var x, y int = 1, 2
 	println(add(x, y))
 }
+
+// Error:
+// 10:11: int does not implement main.Float
